Move examples src
