+++ .gitignore
@@ -2,7 +2,7 @@
 /coverage/*
 /build/*
 /dist/*
-/examples/__build__/*
+/examples/build/*
 /.vscode/*
 
 npm-debug.log*
