+++ examples/index.html
@@ -14,5 +14,5 @@
   <div id="custom-menu"></div>
   <div id="managed-menu-visibility"></div>
 
-  <script src="bundle.js"></script>
+  <script src="build/bundle.js"></script>
 </body>
