+++ package.json
@@ -4,7 +4,7 @@
   "description": "Accessible, extensible, Autocomplete for React.js",
   "main": "./build/lib/Autocomplete.js",
   "scripts": {
-    "start": "budo examples/index.tsx:bundle.js --live --dir examples -- -p [tsify -p tsconfig.examples.json]",
+    "start": "budo examples/src/index.tsx:build/bundle.js --live --dir examples -- -p [tsify -p tsconfig.examples.json]",
     "prettier": "prettier --write '**/*.{ts,tsx,js,jsx,html,css,md,mk,json}'",
     "prettier-check": "prettier --check '**/*.{ts,tsx,js,jsx,html,css,md,mk,json}'",
     "lint": "eslint '**/*.{ts,tsx,js,jsx}'",
